+++ js-sdk/packages/content/test/index.spec.ts
@@ -67,7 +67,8 @@ describe('Engine Client', () => {
 					.get(endpoints.GET_ITEM_URL)
 					.query({
 						crafterSite,
-						url: '/site/website/index.xml'
+						url: '/site/website/index.xml',
+						flatten: false
 					})
 					.reply(200, item);
 
@@ -84,26 +85,6 @@ describe('Engine Client', () => {
 			});
 		});
 
-		describe('getDescriptor', () => {
-			// Tests the contentStore getDescriptor method. Checks that it returns the index descriptor of the site editorial.
-			it('return the index descriptor', (done) => {
-				nock(baseUrl)
-					.get(endpoints.GET_DESCRIPTOR)
-					.query({
-						crafterSite,
-						flatten: false,
-						url: '/site/website/index.xml'
-					})
-					.reply(200, descriptor);
-
-				ContentStoreService.getDescriptor('/site/website/index.xml').subscribe((respDescriptor) => {
-					expect(respDescriptor).to.not.be.null;
-					expect(respDescriptor.page.objectId).to.equal(descriptor.page.objectId);
-					done();
-				});
-			});
-		});
-
 		describe('getChildren', () => {
 			// Tests the contentStore getChildren method. Checks that it returns the children of the index page of the site editorial.
 			// The children length at its ids should match the expected values.
@@ -112,7 +93,8 @@ describe('Engine Client', () => {
 					.get(endpoints.GET_CHILDREN)
 					.query({
 						crafterSite,
-						url: '/site/website/'
+						url: '/site/website/',
+						flatten: false
 					})
 					.reply(200, children);
 
@@ -146,7 +128,8 @@ describe('Engine Client', () => {
 					.query({
 						crafterSite,
 						depth: 3,
-						url: '/site/website/articles/2021'
+						url: '/site/website/articles/2021',
+						flatten: false
 					})
 					.reply(200, tree);